Adding a workable version of hashSet.
However, the implementation it is only an ack to have a workable implementation.

+++ go/datastructures/set/treeset/tree_set_test.go
@@ -8,5 +8,5 @@ import (
 
 func TestTreeSetInt(t *testing.T) {
 	set := NewTreeSet()
-	assert.Equal(t, uint64(0), set.Size())
+	assert.Equal(t, 0, set.Size())
 }
